+++ Content/Enemies/maneater.cs
@@ -21,7 +21,7 @@ namespace MurphysMod.Content.Enemies
                 {
                     Player player = Main.LocalPlayer;
                     LuckHandler luckHandler = player.GetModPlayer<LuckHandler>();
-                    float luckVal = luckHandler.luckValue();
+                    double luckVal = luckHandler.luckValue();
                     var knockback = 2f;
                     int x = 0;
