+++ Content/Enemies/worm.cs
@@ -22,7 +22,7 @@ namespace MurphysMod.Content.Enemies
 
                 Player player = Main.LocalPlayer;
                 LuckHandler luckHandler = player.GetModPlayer<LuckHandler>();
-                float luckVal = luckHandler.luckValue();
+                double luckVal = luckHandler.luckValue();
                 var knockback = 2f;
                 int x = 0;
