+++ Content/Projectiles/MagmaKnifeProjectile.cs
@@ -31,7 +31,7 @@ namespace MurphysMod.Content.Enemies
         {
             Player player = Main.LocalPlayer;
             LuckHandler luckHandler = player.GetModPlayer<LuckHandler>();
-            float luckVal = luckHandler.luckValue();
+            double luckVal = luckHandler.luckValue();
 
             if (luckVal >= .25f)
             {
